requirements.txt
torch
numpy
coqui-tts
jiwer
transformers
sentence-transformers
brisque
Pillow
pathlib
openai
requests
moviepy